backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.8.0
django-nose==1.4.7
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
lxml==4.9.1
MarkupSafe==2.0.1
nose==1.3.7
oauthlib==3.1.1
openapi-codec==1.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.6
django-smtp-ssl
django-apscheduler==0.6.2
APScheduler~=3.9.1
ua-parser==0.10.0
user-agents==2.2.0
pytube==12.0.0
drf-yasg2==1.19.4
google-api-core
google-auth
google-cloud-core
google-cloud-translate
